refactor: move repository to root package
