+++ app/src/androidTest/java/dev/mslalith/focuslauncher/data/repository/AppDrawerRepoTest.kt
@@ -7,7 +7,6 @@ import app.cash.turbine.test
 import com.google.common.truth.Truth.assertThat
 import dev.mslalith.focuslauncher.TestApps
 import dev.mslalith.focuslauncher.data.database.AppDatabase
-import dev.mslalith.focuslauncher.data.network.repository.AppDrawerRepo
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runTest
