+++ app/src/androidTest/java/dev/mslalith/focuslauncher/data/repository/ClockRepoTest.kt
@@ -1,7 +1,6 @@
 package dev.mslalith.focuslauncher.data.repository
 
 import com.google.common.truth.Truth.assertThat
-import dev.mslalith.focuslauncher.data.network.repository.ClockRepo
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
 import org.junit.After
