+++ app/src/androidTest/java/dev/mslalith/focuslauncher/data/repository/settings/AppDrawerSettingsRepoTest.kt
@@ -2,7 +2,6 @@ package dev.mslalith.focuslauncher.data.repository.settings
 
 import com.google.common.truth.Truth.assertThat
 import dev.mslalith.focuslauncher.data.model.AppDrawerViewType
-import dev.mslalith.focuslauncher.data.network.repository.settings.AppDrawerSettingsRepo
 import dev.mslalith.focuslauncher.data.repository.DataStoreTest
 import dev.mslalith.focuslauncher.data.utils.Constants
 import kotlinx.coroutines.ExperimentalCoroutinesApi
