+++ app/src/androidTest/java/dev/mslalith/focuslauncher/data/repository/settings/LunarPhaseSettingsRepoTest.kt
@@ -1,7 +1,6 @@
 package dev.mslalith.focuslauncher.data.repository.settings
 
 import com.google.common.truth.Truth.assertThat
-import dev.mslalith.focuslauncher.data.network.repository.settings.LunarPhaseSettingsRepo
 import dev.mslalith.focuslauncher.data.repository.DataStoreTest
 import dev.mslalith.focuslauncher.data.utils.Constants
 import kotlinx.coroutines.ExperimentalCoroutinesApi
