+++ app/src/androidTest/java/dev/mslalith/focuslauncher/data/repository/settings/QuotesSettingsRepoTest.kt
@@ -1,7 +1,6 @@
 package dev.mslalith.focuslauncher.data.repository.settings
 
 import com.google.common.truth.Truth.assertThat
-import dev.mslalith.focuslauncher.data.network.repository.settings.QuotesSettingsRepo
 import dev.mslalith.focuslauncher.data.repository.DataStoreTest
 import dev.mslalith.focuslauncher.data.utils.Constants
 import kotlinx.coroutines.ExperimentalCoroutinesApi
